jpms: Refactor handlebars.guava
