+++ handlebars-guava-cache/src/main/java/com/github/jknack/handlebars/io/guava/GuavaCachedTemplateLoader.java
@@ -15,7 +15,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.github.jknack.handlebars.io;
+package com.github.jknack.handlebars.io.guava;
 
 import java.io.IOException;
 import java.nio.charset.Charset;
@@ -23,6 +23,8 @@ import java.util.concurrent.ExecutionException;
 import java.util.concurrent.TimeUnit;
 
 import com.github.jknack.handlebars.Handlebars;
+import com.github.jknack.handlebars.io.TemplateLoader;
+import com.github.jknack.handlebars.io.TemplateSource;
 import com.google.common.cache.Cache;
 import com.google.common.cache.CacheBuilder;
